pyproject.toml
[project]
authors = [
    { name = "KimigaiiWuyi", email = "444835641@qq.com" },
    { name = "MingxuanGame", email = "MingxuanGame@outlook.com" },
]
license = { text = "GPL-3.0-or-later" }
requires-python = ">=3.11,<4.0"
dependencies = [
    "httpx>=0.23.0",
    "beautifulsoup4>=4.11.1",
    "lxml>=4.9.2",
    "aiohttp>=3.8.1",
    "sqlalchemy<2.1.0,>=2.0.0",
    "pillow>=10.2.0",
    "aiosqlite>=0.17.0",
    "aiofiles>=0.8.0",
    "gitpython>=3.1.27",
    "qrcode[pil]<8.0.0,>=7.3.1",
    "msgspec>=0.13.1",
    "uvicorn>=0.20.0",
    "websockets>=13.1",
    "urllib3<2.0.0,>=1.26.15",
    "mpmath<2.0.0,>=1.3.0",
    "fastapi>=0.97.0,<0.112.3",
    "apscheduler<4.0.0,>=3.10.1",
    "aioboto3<13.0.0,>=12.0.0",
    "jinja2<4.0.0,>=3.1.2",
    "colorama<1.0.0,>=0.4.6",
    "async-timeout>=4.0.3",
    "toml>=0.10.2",
    "bcrypt==4.0.1",
    "setuptools>=69.0.3",
    "email-validator>=2.2.0",
    "annotated-types>=0.7.0",
    "sqlmodel<0.0.31",
    "fastapi<0.112.3",
    "async-timeout<5.0.0,>=4.0.3",
    "toml<1.0.0,>=0.10.2",
    "bcrypt<5.0.0,>=4.0.1",
    "email-validator<3.0.0,>=2.2.0",
    "psutil>=6.1.1",
    "pytz>=2024.2",
    "pandas>=2.2.3",
    "betterproto>=2.0.0b7",
    "structlog>=25.3.0",
    "rich>=14.0.0",
    "cryptography>=45.0.5",
    "packaging>=25.0",
    "greenlet>=3.2.4",
    "boltons>=25.0.0",
    "lmdb>=1.7.5",
    "openai>=2.15.0",
    "google-generativeai>=0.8.6",
    "jieba>=0.42.1",
    "joblib>=1.5.3",
    "scikit-learn>=1.7.2",
    "pydantic>=2.12.2",
    "pydantic-settings>=2.12.0",
    "pydantic-core>=2.20.1",
    "playwright>=1.58.0",
    "fastembed>=0.7.4",
    "qdrant-client>=1.17.0",
    "tiktoken>=0.12.0",
    "python-multipart>=0.0.22",
    "tavily-python>=0.7.23",
    "pydantic-ai-slim[anthropic,openai]>=1.72.0",
    "pydantic-ai-skills>=0.6.0",
    "json-repair>=0.59.4",
]
name = "gsuid-core"
version = "0.9.5"
description = "KimigaiiWuyi/GenshinUID 的核心部分，平台无关，便于移植到其他平台以及框架。"
readme = "README.md"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.group.dev.dependencies]
ruff = "^0.14.7"
pre-commit = "^2.21.0"
pytest = "^8.3.0"

[tool.pdm.scripts]
core = "gsuid_core/core.py"

[tool.pdm.dev-dependencies]
dev = ["ruff>=0.14.7", "pre-commit>=2.21.0", "pytest>=8.3.0"]

[tool.pdm.build]
includes = ["gsuid_core"]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[dependency-groups]
dev = ["ruff>=0.14.7", "pre-commit>=2.21.0", "pytest>=8.3.0"]

[tool.uv]
default-groups = []

[tool.pdm.resolution]
respect-source-order = true


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[project.urls]
"Bug Tracker" = "https://github.com/Genshin-bots/gsuid-core/issues"
homepage = "https://github.com/Genshin-bots/gsuid-core"
repository = "https://github.com/Genshin-bots/gsuid-core"

[project.scripts]
core = "gsuid_core.core:main"
